requirements.txt
pandas
numpy
faker
cx_Oracle
sqlalchemy
python-dotenv
scikit-learn
xgboost
shap
matplotlib
seaborn
plotly